Revert "fix intel compiler warning"
This reverts commit 87451328e9046349d6424985e91064e5b4411f93.

+++ src/scip/cons_expr_abs.c
@@ -484,7 +484,7 @@ SCIP_DECL_CONSEXPR_EXPRHASH(hashAbs)
    assert(expr2key != NULL);
    assert(hashkey != NULL);
 
-   *hashkey = (unsigned int)ABS_HASHKEY;
+   *hashkey = ABS_HASHKEY;
 
    assert(SCIPhashmapExists(expr2key, (void*) SCIPgetConsExprExprChildren(expr)[0]));
    childhash = (unsigned int)(size_t) SCIPhashmapGetImage(expr2key, SCIPgetConsExprExprChildren(expr)[0]);
